requirements.txt
psycopg2-binary
pymongo
sentry-sdk
mwclient
requests